--- src/components/index.js
@@ -1,4 +1,0 @@
-export { default as Header } from './Header.js'
-export { default as Navigation } from './Navigation.js'
-export { default as Footer } from './Footer.js'
-export { default as Layout } from './Layout.js'
