+++ src/components/layouts/MainLayout.js
@@ -1,6 +1,6 @@
-import { Navigation, Header, Footer } from '@components'
+import { Navigation, Header, Footer } from '@components/layouts'
 
-export default function Layout(children) {
+export default function MainLayout(children) {
   return `<div class="bg-gray-100 min-h-screen flex justify-center">
     <div class="max-w-md w-full">
       ${Header()}
