+++ src/components/main/index.js
@@ -1,0 +1,1 @@
+export { default as FeedItem } from './FeedItem'
